created modules for javascript file
