+++ README.md
@@ -1,3 +1,7 @@
 # fake-OS
 
 Fake OS is a single page application made to look like a Windows 98 Operating System which has 3 clickable applications that all function correctly using modals.
+
+The application has a clickable start button, with options that change on hover.
+
+The application also has 3 modal based apps including a music player, calculator, and wallpaper changer.
