+++ index.html
@@ -3,7 +3,11 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
+    <script type="module" src="./script.js" defer></script>
+    <script type="module" src="./calculator.js" defer></script>
+    <script type="module" src="./music-player.js" defer></script>
+    <script type="module" src="./time.js"></script>
+    <script type="module" src="./change-wallpaper.js"></script>
     <link rel="stylesheet" href="style.css" />
     <title>Fake OS</title>
   </head>
